Extract arc/arcLine/connectingLine packers as pure functions
Same treatment as read/cigar packers — pull the inner upload loops
out of the class into pure pack*() helpers with hoisted locals. The
upload methods on the renderer drop to plain dispatch.

+++ plugins/variants/src/MultiVariantDisplay/components/GpuVariantRenderer.ts
@@ -46,7 +46,7 @@ export class GpuVariantRenderer implements VariantBackend {
       regionStart: number
       cellPositions: Uint32Array
       cellRowIndices: Uint32Array
-      cellColors: Uint8Array
+      cellColors: Uint32Array
       cellShapeTypes: Uint8Array
       numCells: number
     },
